Migrate user related plan
