+++ src/query/legacy-planners/src/lib.rs
@@ -55,9 +55,6 @@ mod plan_sink;
 mod plan_sort;
 mod plan_subqueries_set;
 mod plan_table_recluster;
-mod plan_user_alter;
-mod plan_user_create;
-mod plan_user_drop;
 mod plan_user_stage_create;
 mod plan_user_stage_describe;
 mod plan_user_stage_drop;
@@ -152,9 +149,6 @@ pub use plan_sink::SINK_SCHEMA;
 pub use plan_sort::SortPlan;
 pub use plan_subqueries_set::SubQueriesSetPlan;
 pub use plan_table_recluster::ReclusterTablePlan;
-pub use plan_user_alter::AlterUserPlan;
-pub use plan_user_create::CreateUserPlan;
-pub use plan_user_drop::DropUserPlan;
 pub use plan_user_stage_create::CreateUserStagePlan;
 pub use plan_user_stage_describe::DescribeUserStagePlan;
 pub use plan_user_stage_drop::DropUserStagePlan;
