+++ src/query/planner/src/plans/drop_user.rs
@@ -1,4 +1,4 @@
-// Copyright 2021 Datafuse Labs.
+// Copyright 2022 Datafuse Labs.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -11,24 +11,19 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 use std::sync::Arc;
 
 use common_datavalues::DataSchema;
 use common_datavalues::DataSchemaRef;
-use common_meta_types::AuthInfo;
 use common_meta_types::UserIdentity;
-use common_meta_types::UserOption;
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug, PartialEq, Eq)]
-pub struct CreateUserPlan {
+pub struct DropUserPlan {
+    pub if_exists: bool,
     pub user: UserIdentity,
-    pub auth_info: AuthInfo,
-    pub user_option: UserOption,
-    pub if_not_exists: bool,
 }
 
-impl CreateUserPlan {
+impl DropUserPlan {
     pub fn schema(&self) -> DataSchemaRef {
         Arc::new(DataSchema::empty())
     }
