+++ src/query/planner/src/plans/mod.rs
@@ -13,14 +13,17 @@
 // limitations under the License.
 
 mod alter_table_cluster_key;
+mod alter_user;
 mod call;
 mod create_database;
 mod create_role;
+mod create_user;
 mod describe_table;
 mod drop_database;
 mod drop_role;
 mod drop_table;
 mod drop_table_cluster_key;
+mod drop_user;
 mod exists_table;
 mod grant_privilege;
 mod grant_role;
@@ -40,14 +43,17 @@ mod undrop_table;
 mod use_database;
 
 pub use alter_table_cluster_key::AlterTableClusterKeyPlan;
+pub use alter_user::AlterUserPlan;
 pub use call::CallPlan;
 pub use create_database::CreateDatabasePlan;
 pub use create_role::CreateRolePlan;
+pub use create_user::CreateUserPlan;
 pub use describe_table::DescribeTablePlan;
 pub use drop_database::DropDatabasePlan;
 pub use drop_role::DropRolePlan;
 pub use drop_table::DropTablePlan;
 pub use drop_table_cluster_key::DropTableClusterKeyPlan;
+pub use drop_user::DropUserPlan;
 pub use exists_table::ExistsTablePlan;
 pub use grant_privilege::GrantPrivilegePlan;
 pub use grant_role::GrantRolePlan;
