+++ src/query/service/src/sql/planner/binder/ddl/account.rs
@@ -19,12 +19,12 @@ use common_ast::ast::CreateUserStmt;
 use common_ast::ast::GrantStmt;
 use common_ast::ast::RevokeStmt;
 use common_exception::Result;
-use common_legacy_planners::AlterUserPlan;
-use common_legacy_planners::CreateUserPlan;
 use common_meta_types::AuthInfo;
 use common_meta_types::GrantObject;
 use common_meta_types::UserOption;
 use common_meta_types::UserPrivilegeSet;
+use common_planner::plans::AlterUserPlan;
+use common_planner::plans::CreateUserPlan;
 use common_planner::plans::GrantPrivilegePlan;
 use common_planner::plans::GrantRolePlan;
 use common_planner::plans::RevokePrivilegePlan;
