+++ src/query/service/src/sql/planner/binder/mod.rs
@@ -27,13 +27,13 @@ use common_datavalues::DataTypeImpl;
 use common_exception::Result;
 use common_legacy_planners::AlterUserUDFPlan;
 use common_legacy_planners::CreateUserUDFPlan;
-use common_legacy_planners::DropUserPlan;
 use common_legacy_planners::DropUserStagePlan;
 use common_legacy_planners::DropUserUDFPlan;
 use common_meta_types::UserDefinedFunction;
 use common_planner::plans::CallPlan;
 use common_planner::plans::CreateRolePlan;
 use common_planner::plans::DropRolePlan;
+use common_planner::plans::DropUserPlan;
 use common_planner::plans::ShowGrantsPlan;
 use common_planner::plans::UseDatabasePlan;
 use common_planner::MetadataRef;
