+++ src/query/service/src/sql/planner/plans/mod.rs
@@ -44,15 +44,12 @@ use common_datavalues::DataSchema;
 use common_datavalues::DataSchemaRef;
 use common_datavalues::DataSchemaRefExt;
 use common_datavalues::StringType;
-use common_legacy_planners::AlterUserPlan;
 use common_legacy_planners::AlterUserUDFPlan;
 use common_legacy_planners::AlterViewPlan;
-use common_legacy_planners::CreateUserPlan;
 use common_legacy_planners::CreateUserStagePlan;
 use common_legacy_planners::CreateUserUDFPlan;
 use common_legacy_planners::CreateViewPlan;
 use common_legacy_planners::DeletePlan;
-use common_legacy_planners::DropUserPlan;
 use common_legacy_planners::DropUserStagePlan;
 use common_legacy_planners::DropUserUDFPlan;
 use common_legacy_planners::DropViewPlan;
@@ -60,14 +57,17 @@ use common_legacy_planners::ReclusterTablePlan;
 use common_legacy_planners::RemoveUserStagePlan;
 use common_legacy_planners::SettingPlan;
 use common_planner::plans::AlterTableClusterKeyPlan;
+use common_planner::plans::AlterUserPlan;
 use common_planner::plans::CallPlan;
 use common_planner::plans::CreateDatabasePlan;
 use common_planner::plans::CreateRolePlan;
+use common_planner::plans::CreateUserPlan;
 use common_planner::plans::DescribeTablePlan;
 use common_planner::plans::DropDatabasePlan;
 use common_planner::plans::DropRolePlan;
 use common_planner::plans::DropTableClusterKeyPlan;
 use common_planner::plans::DropTablePlan;
+use common_planner::plans::DropUserPlan;
 use common_planner::plans::ExistsTablePlan;
 use common_planner::plans::GrantPrivilegePlan;
 use common_planner::plans::GrantRolePlan;
